Clickhouse fix (#4971)
* fix production build after webpack.config changes
