+++ webpack.config.js
@@ -45,6 +45,7 @@ const webpack = require('webpack');
 const path = require('path');
 const { watch } = require('fs');
 const fs = require('fs/promises');
+const { mergeApiTestJson } = require('./apiTests/mergeJson');
 
 const join = path.join;
 const resolve = path.resolve;
@@ -597,36 +598,14 @@ if (isTest) {
 }
 // End Testing
 
-async function mergeApiTestJson() {
-    const files = (await fsProm.readdir('./apiTests/specs')).map(fileName => {
-        return path.join('./apiTests/specs', fileName);
-    });
-
-    const jsons = files.map(path => {
-        return fsProm.readFile(path).then(data => {
-            try {
-                const json = JSON.parse(data);
-                return { file: path, suites: json };
-            } catch (ex) {
-                console.log('invalid apiTest json spec');
-                return [];
-            }
-        });
-    });
-
-    Promise.all(jsons)
-        .then(d => {
-            fsProm.writeFile('./apiTests/merged-tests.json', JSON.stringify(d));
-        })
-        .then(r => console.log('merged-tests.json written'));
-}
-
 mergeApiTestJson();
 
-watch('./apiTests/specs', async function(event, filename) {
-    if (event === 'change') {
-        mergeApiTestJson();
-    }
-});
+if (isDev) {
+    watch('./apiTests/specs', async function(event, filename) {
+        if (event === 'change') {
+            mergeApiTestJson();
+        }
+    });
+}
 
 module.exports = config;
